Move the lobby to the top northeast corner
Reserve a cube with roughly 300 meter side (to be exact, it
is 4x4x4 mapchunk cube) at the top northeast corner for game
start stuff. Right now the lobby is located there and the
rest is filled with bedrock.

This was done to make sure the mapgen does not get invoked
by the player presence for any of the ground stuff. The
mapgen needs to be managed by the code to make sure the
mapchunks are generated in the correct order.