Do not record duplicated emerges
Minetest 0.4.16 has a bug where mapgen can be invoked for a
mapchunk more than once. This piece of code keeps track of
which chunks were already seen by the mapgen and skips the
LUA portion for those that were seen previously. The code
that records the chunks is after the check so as a
consequence, duplicated chunks are reported only when they
were first encountered.

This also introduces an utility function which iterates over
lines of a text file, calling user specified callback on each
line.