High resolution clock
This clock counts in tenths of seconds. That gives more
resolution than the minetest.get_gametime() which returns
time in whole seconds only.

The clock is implemented using minetest.after but a somewhat
involved block of code is included to make sure the clock
does not lag too much begind the real minetest.get_gametime()

As a precaution, minetest.get_gametime() was removed from the
API. Use base.GetTime() to get the time.