Fix periodic hook crash in LuaJIT
In LuaJIT variadic argument handling is broken: the arg is
sometimes set to "nil" when a variadic procedure receives no
arguments and if it is not set to "nil", it is set to an
empty table even when there are arguments. This makes
variadic arguments completely unusable so refactor the code
to use single argument in these cases instead.